Parameter to replace label with image
