+++ tests/test_save.py
@@ -34,8 +34,7 @@ from pydicom.uid import UID, JPEGBaseline8Bit, generate_uid
 from wsidicom import WsiDicom
 from wsidicom.geometry import Point, Size, SizeMm
 from wsidicom.image_data import ImageData, ImageOrigin
-from wsidicom.file import WsiDicomFile
-from wsidicom.file_writer import WsiDicomFileWriter
+from wsidicom.file import WsiDicomFile, WsiDicomFileWriter
 from wsidicom.uid import WSI_SOP_CLASS_UID
 from wsidicom.wsidicom import WsiDicomLevel
