+++ wsidicom/instance.py
@@ -1,4 +1,4 @@
-#    Copyright 2021, 2022 SECTRA AB
+#    Copyright 2021, 2022, 2023 SECTRA AB
 #
 #    Licensed under the Apache License, Version 2.0 (the "License");
 #    you may not use this file except in compliance with the License.
@@ -20,16 +20,18 @@ from typing import (Callable, DefaultDict, Dict, List, Optional, OrderedDict,
                     Sequence, Set, Tuple, Union, cast)
 
 from PIL import Image
+from pydicom import Dataset
 from pydicom.uid import UID
 
-from wsidicom.dataset import WsiDataset
+from wsidicom.dataset import ImageType, WsiDataset
 from wsidicom.errors import (WsiDicomError, WsiDicomMatchError,
                              WsiDicomNotFoundError, WsiDicomOutOfBoundsError,
                              WsiDicomUidDuplicateError)
-from wsidicom.file import WsiDicomFile
-from wsidicom.file_writer import WsiDicomFileWriter
+from wsidicom.file import WsiDicomFile, WsiDicomFileWriter
 from wsidicom.geometry import Point, Region, RegionMm, Size, SizeMm
-from wsidicom.image_data import ImageData, ImageOrigin, WsiDicomImageData
+from wsidicom.image_data import ImageData, ImageOrigin
+from wsidicom.image_data.dicom_image_data import WsiDicomImageData
+from wsidicom.image_data.pillow_image_data import PillowImageData
 from wsidicom.stringprinting import dict_pretty_str
 from wsidicom.uid import SlideUids
 
@@ -56,7 +58,7 @@ class WsiInstance:
         self._datasets = datasets
         self._image_data = image_data
         self._identifier, self._uids = self._validate_instance(self.datasets)
-        self._wsi_type = self.dataset.wsi_type
+        self._image_type = self.dataset.image_type
 
         if self.ext_depth_of_field:
             if self.ext_depth_of_field_planes is None:
@@ -91,9 +93,9 @@ class WsiInstance:
         return string
 
     @property
-    def wsi_type(self) -> str:
+    def image_type(self) -> ImageType:
         """Return wsi type."""
-        return self._wsi_type
+        return self._image_type
 
     @property
     def datasets(self) -> Sequence[WsiDataset]:
@@ -118,12 +120,14 @@ class WsiInstance:
         return self._image_data.tile_size
 
     @property
-    def mpp(self) -> SizeMm:
+    def mpp(self) -> Optional[SizeMm]:
         """Return pixel spacing in um/pixel."""
+        if self.pixel_spacing is None:
+            return None
         return self.pixel_spacing*1000.0
 
     @property
-    def pixel_spacing(self) -> SizeMm:
+    def pixel_spacing(self) -> Optional[SizeMm]:
         """Return pixel spacing in mm/pixel."""
         return self._image_data.pixel_spacing
 
@@ -240,6 +244,64 @@ class WsiInstance:
             for instance_files in files_grouped_by_instance.values()
         ]
 
+    @classmethod
+    def create_label(
+        cls,
+        image: Union[Image.Image, str, Path],
+        base_dataset: Dataset,
+        instance_number: int
+        # TODO Should not need to specify instance number
+        # Should be set on write.
+    ) -> 'WsiInstance':
+        if isinstance(image, Image.Image):
+            image_data = PillowImageData(image)
+        else:
+            image_data = PillowImageData.from_file(image)
+        return cls.create_instance(
+            image_data,
+            base_dataset,
+            ImageType.LABEL,
+            instance_number
+        )
+
+    @classmethod
+    def create_instance(
+        cls,
+        image_data: ImageData,
+        base_dataset: Dataset,
+        image_type: ImageType,
+        instance_number: int
+    ) -> 'WsiInstance':
+        """Create WsiInstance from ImageData.
+
+        Parameters
+        ----------
+        image_data: ImageData
+            Image data and metadata.
+        base_dataset: Dataset
+            Base dataset to include.
+        image_type: ImageType
+            Type of instance to create.
+        instance_number: int
+            The number of the instance (in a series).
+
+        Returns
+        ----------
+        WsiInstance
+            Created WsiInstance.
+        """
+        instance_dataset = WsiDataset.create_instance_dataset(
+            base_dataset,
+            image_type,
+            instance_number,
+            image_data
+        )
+
+        return cls(
+            instance_dataset,
+            image_data
+        )
+
     @staticmethod
     def check_duplicate_instance(
         instances: Sequence['WsiInstance'],
@@ -307,7 +369,7 @@ class WsiInstance:
             self.uids.matches(other_instance.uids) and
             self.size == other_instance.size and
             self.tile_size == other_instance.tile_size and
-            self.wsi_type == other_instance.wsi_type
+            self.image_type == other_instance.image_type
         )
 
     def close(self) -> None:
@@ -335,7 +397,7 @@ class WsiDicomGroup:
         self._validate_group()
 
         base_instance = instances[0]
-        self._wsi_type = base_instance.wsi_type
+        self._image_type = base_instance.image_type
         self._uids = base_instance.uids
 
         self._size = base_instance.size
@@ -374,9 +436,9 @@ class WsiDicomGroup:
         return self._uids
 
     @property
-    def wsi_type(self) -> str:
+    def image_type(self) -> ImageType:
         """Return wsi type"""
-        return self._wsi_type
+        return self._image_type
 
     @property
     def size(self) -> Size:
@@ -384,12 +446,14 @@ class WsiDicomGroup:
         return self._size
 
     @property
-    def mpp(self) -> SizeMm:
+    def mpp(self) -> Optional[SizeMm]:
         """Return pixel spacing in um/pixel"""
+        if self.pixel_spacing is None:
+            return None
         return self.pixel_spacing*1000.0
 
     @property
-    def pixel_spacing(self) -> SizeMm:
+    def pixel_spacing(self) -> Optional[SizeMm]:
         """Return pixel spacing in mm/pixel"""
         return self._pixel_spacing
 
@@ -484,7 +548,7 @@ class WsiDicomGroup:
         """
         return (
             self.uids.matches(other_group.uids) and
-            other_group.wsi_type == self.wsi_type
+            other_group.image_type == self.image_type
         )
 
     def valid_pixels(self, region: Region) -> bool:
@@ -718,6 +782,8 @@ class WsiDicomGroup:
         Region
             Region in pixels
         """
+        if self.pixel_spacing is None:
+            raise Exception()  # TODO
         pixel_region = Region(
             position=region.position // self.pixel_spacing,
             size=region.size // self.pixel_spacing
@@ -998,6 +1064,18 @@ class WsiDicomLevel(WsiDicomGroup):
         """Return pyramid level"""
         return self._level
 
+    @property
+    def mpp(self) -> SizeMm:
+        if self.pixel_spacing is None:
+            raise Exception()  # TODO
+        return self.pixel_spacing*1000.0
+
+    @property
+    def pixel_spacing(self) -> SizeMm:
+        if self._pixel_spacing is None:
+            raise Exception()  # TODO
+        return self._pixel_spacing
+
     @classmethod
     def open(
         cls,
@@ -1020,6 +1098,8 @@ class WsiDicomLevel(WsiDicomGroup):
         instances_grouped_by_level = cls._group_instances(instances)
         base_group = list(instances_grouped_by_level.values())[0]
         base_pixel_spacing = base_group[0].pixel_spacing
+        if base_pixel_spacing is None:
+            raise Exception()  # TODO
         for level in instances_grouped_by_level.values():
             levels.append(cls(level, base_pixel_spacing))
         return levels
@@ -1041,7 +1121,7 @@ class WsiDicomLevel(WsiDicomGroup):
         other_level = cast(WsiDicomLevel, other_level)
         return (
             self.uids.matches(other_level.uids) and
-            other_level.wsi_type == self.wsi_type and
+            other_level.image_type == self.image_type and
             other_level.tile_size == self.tile_size
         )
